requirements.txt
# Sentinel-Ops Dependencies

# Cloud Function framework
functions-framework>=3.0.0

# Odoo XML-RPC (built into Python stdlib - xmlrpc.client)

# BigQuery for audit and KPIs
google-cloud-bigquery>=3.0.0

# Secret Manager for production secrets
google-cloud-secret-manager>=2.0.0

# HTTP requests (for Slack alerts)
requests>=2.28.0

# Environment variable loading
python-dotenv>=1.0.0

# Development/Testing
pytest>=7.0.0
pytest-mock>=3.10.0
